backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "torch==2.8.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
    "whisperx>=3.7.2",
    "torchaudio==2.8.0",
    "python-multipart>=0.0.21",
    "ffmpeg-python>=0.2.0",
    "redis>=7.1.0",
]
